Update json io

+++ Game_Logic/json_io.py
@@ -9,7 +9,7 @@ def json_reader(file_name):
     :return:
     """
     if not os.path.isfile(file_name):
-        raise IOError("JSON file does not exist.")
+        raise IOError("JSON file {} does not exist.".format(file_name))
     with open(file_name, 'r', encoding='utf-8') as f:
         data = json.load(f)
     return data
@@ -22,6 +22,8 @@ def json_writer(file_name, data):
     :param data:
     :return:
     """
+    if not os.path.isfile(file_name):
+        raise IOError("JSON file {} does not exist.".format(file_name))
     with open(file_name, 'w') as f:
         json.dump(data, f, indent=4)
